update arith-tokenizer.js
(Summary) Fixed an issue where calling 'Tokenizer.next ()' after reaching the end of the sequence would result in a runtime error.

(Details) In the state of "Tokenizer.hasNext () === false", the return value of "Tokenizer.next ()" must return "undefined". However, the argument of "Tokenizer._asNumber (char)" in the first "if" of the internal process "Tokenizer._getSentence ()" cannot be "undefined". Before that, it was changed to determine "char === undefined".